xcopy v0.1: Initial smoke-tested xcopy
This commit is a pre-RFC (v0.1) smoke-tested version of a splice()
based xcopy(). We used splice to provide a convenient interface to the
kernel and then introduce a new splice flag (SPLICE_F_P2PDMA) in order
to indicate that userspace would like to do a p2pdma based copy.

This repo contains a simple userspace application to test the
xcopy. It also contains a kernel config file for this v0.1 version of
the kernel (based on 5.3-rc7) as well as a QEMU script that allows for
testing by modelling two NVMe SSDs below a PCIe switch.

See the README file for more information. This commit is tagged v0.1.

Signed-off-by: Stephen Bates <sbates@raithlin.com>